requirements.txt
streamlit==1.30.0
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
scikit-image==0.21.0